requirements.txt
jinja2
openai
pydantic
sqlalchemy
networkx
fastapi
uvicorn
python-dotenv
dashscope
paddlepaddle
paddleocr
pyjwt
python-jose
numpy==1.26.4
pandas==1.5.3
passlib==1.7.4
bcrypt==3.2.0
python-multipart
httpx
pytest
Pillow
pydantic-settings
celery
redis
chromadb
slowapi
